move Dart to own dir
