+++ packages/quicktype-core/src/language/Dart/constants.ts
@@ -1,0 +1,71 @@
+export const keywords = [
+    "abstract",
+    "do",
+    "import",
+    "super",
+    "as",
+    "dynamic",
+    "in",
+    "switch",
+    "assert",
+    "else",
+    "interface",
+    "sync*",
+    "async",
+    "enum",
+    "is",
+    "this",
+    "async*",
+    "export",
+    "library",
+    "throw",
+    "await",
+    "external",
+    "mixin",
+    "true",
+    "break",
+    "extends",
+    "new",
+    "try",
+    "case",
+    "factory",
+    "null",
+    "typedef",
+    "catch",
+    "false",
+    "operator",
+    "var",
+    "class",
+    "final",
+    "part",
+    "void",
+    "const",
+    "finally",
+    "rethrow",
+    "while",
+    "continue",
+    "for",
+    "return",
+    "with",
+    "covariant",
+    "get",
+    "set",
+    "yield",
+    "default",
+    "if",
+    "static",
+    "yield*",
+    "deferred",
+    "implements",
+    "int",
+    "double",
+    "bool",
+    "Map",
+    "List",
+    "String",
+    "File",
+    "fromJson",
+    "toJson",
+    "fromMap",
+    "toMap"
+] as const;
