+++ packages/quicktype-core/src/language/Dart/index.ts
@@ -1,0 +1,2 @@
+export { DartTargetLanguage, dartOptions } from "./language";
+export { DartRenderer } from "./DartRenderer";
